增加Job修改时校验alist和本地目录是否一致 抽取了jobUtils处理MD5加密和本地目录校验 加密key的计算从文件提到提到了job维度，减少计算
