+++ service/encrypt/encrypt.py
@@ -1,19 +1,17 @@
 import os
-from gmssl.sm4 import CryptSM4, SM4_ENCRYPT, SM4_DECRYPT
+from gmssl.sm4 import CryptSM4, SM4_ENCRYPT
 
 
 class SM4FileCrypto:
 
-    def __init__(self, key):
+    def __init__(self, key,iv):
         """
         初始化SM4文件加密器
         :param key: 16字节密钥
         :param iv: 16字节初始化向量，None则随机生成
         """
-        if len(key) != 32:
-            raise ValueError("密码格式不对")
-        self.key = key[6:22].upper().encode('utf-8')
-        self.iv = key[16:32].upper().encode('utf-8')
+        self.key = key
+        self.iv = iv
         self.cipher = CryptSM4()
         self.cipher.set_key(self.key, SM4_ENCRYPT)
         self.block_size = 16  # SM4分组大小
@@ -26,16 +24,13 @@ class SM4FileCrypto:
         self.counter += 1
         return self.cipher.crypt_ecb(ctr_block)
 
-    def encrypt_file(self, input_file, output_file, chunk_size=8*1024 * 1024):
+    def encrypt_file(self, input_file, output_file, chunk_size=8 * 1024 * 1024):
         """
         加密文件
         :param input_file: 输入文件路径
         :param output_file: 输出文件路径
         :param chunk_size: 每次处理的块大小(字节)
         """
-        print(f'enptkey:{self.key}')
-        print(f'input_file:{input_file}')
-        print(f'output_file:{output_file}')
 
         with open(input_file, 'rb') as fin, open(output_file, 'wb') as fout:
             # 首先写入IV，解密时需要
@@ -55,11 +50,8 @@ class SM4FileCrypto:
 
                 fout.write(encrypted_chunk)
 
-    def decrypt_file(self, input_file, output_file, chunk_size=8*1024 * 1024):
+    def decrypt_file(self, input_file, output_file, chunk_size=8 * 1024 * 1024):
 
-        print(f'enptkey:{self.key}')
-        print(f'input_file:{input_file}')
-        print(f'output_file:{output_file}')
 
         """
         解密文件(加密和解密过程相同)
