+++ service/syncJob/jobClient.py
@@ -4,12 +4,10 @@
 """
 import json
 import logging
-import os.path
 import threading
 import time
 
 from apscheduler.schedulers.background import BackgroundScheduler
-from hashlib import md5
 
 from common.LNG import G
 from common.config import getConfig
@@ -17,6 +15,7 @@ from mapper import jobMapper
 from service.alist import alistSync, alistService
 from service.alist.alistService import getClientById
 from service.syncJob import taskService
+from service.syncJob.jobUtills import alistAndLocalPathMatch, getMd5
 
 
 class JobTask:
@@ -141,17 +140,6 @@ class JobTask:
         taskService.updateJobTaskStatus(self.taskId, status)
 
 
-class utills:
-    # pavel 20250422 密码MD5加密
-    def __init__(self):
-        self.job = None
-
-    def getMd5(string):
-        salt = 'afea'
-        encodeStr = str(string) + salt
-        obj = md5()
-        obj.update(encodeStr.encode("utf-8"))
-        return obj.hexdigest()
 
 
 
@@ -171,45 +159,13 @@ class JobClient:
         if 'method' not in job:
             job['method'] = 0
         if 'id' not in job:
-            # 临时文件夹
-            tempRootFix = '#TEMP_ENPT/'
-
-            #加密和解密场景下，通过创建文件夹在另一端是否能能找到判断配置是否正确
-            #todo 修改时校验待完善
-            encryptFlag = job['encryptFlag']
-            if encryptFlag == 1:
-                alistPath = job['srcPath']
-            elif encryptFlag == 2:
-                alistPath = job['dstPath']
-
-            if encryptFlag != 0:
+            if job['encryptFlag'] != 0:
                 client = getClientById(job['alistId'])
-                localSrcPath = job['localSrcPath']
-
-                dirName = utills.getMd5(localSrcPath)
-
-                alistPath = alistPath + tempRootFix + dirName
-                locaPath = localSrcPath + tempRootFix + dirName
-
-
-                client.mkdir(alistPath)
-
-                flag = os.path.exists(locaPath)
-                if flag:
-                    try:
-                        os.removedirs(locaPath)
-                    except Exception as e:
-                        raise e
-                else:
-                    if encryptFlag == 1:
-                        msg = '本地目录和Alist源目录不匹配'
-                    else:
-                        msg = '本地目录和Alist目标目录不匹配'
-                    raise RuntimeError(msg)
                 # 加密和解密场景下，通过创建文件夹在另一端是否能能找到判断配置是否正确
+                alistAndLocalPathMatch(job, client)
+            #密码加盐加密
+            job['encryptKey'] = getMd5(job['encryptKey'])
 
-                #密码加盐加密
-                job['encryptKey'] = utills.getMd5(job['encryptKey'])
             addJobId = jobMapper.addJob(job)
             job = jobMapper.getJobById(addJobId)
         self.jobId = job['id']
